+++ packages/tealetio/tests/test_streams.py
@@ -15,6 +15,7 @@ from tealetio.streams import (
     SocketTransport,
     StreamReader,
     StreamWriter,
+
     open_connection,
     open_streams,
     run_coro,